[#3455] Modify ldb Tool to be Encryption Aware
Summary: Add `--key_file` option to ldb to pass in a `<key_id>:</path/to/key/file>` and an option `--only_verify_checksums` flag to `scan` to not print out entries and only do checksum validation. Also provide a yb-admin `write_universe_key_to_file` command to retrieve the universe key and write it to a local file.

Test Plan:
Jenkins: build platform: mac
Jenkins: compile only

Reviewers: bogdan, mikhail, timur

Reviewed By: timur

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D7876

+++ src/yb/util/encryption_util.h
@@ -134,7 +134,7 @@ Status CreateEncryptionInfoForWrite(HeaderManager* header_manager,
   // Since file doesn't exist or this overwrites, append key to the name and create.
   *stream = std::make_unique<BlockAccessCipherStream>(std::move(encryption_params));
   RETURN_NOT_OK((*stream)->Init());
-  *header_size = header.size();
+  *header_size = static_cast<uint32_t>(header.size());
   return Status::OK();
 }
